requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
APScheduler==3.10.4
attrs==24.2.0
babel==2.16.0
cachetools==5.5.0
types-cachetools
certifi==2024.8.30
charset-normalizer==3.3.2
environs==11.0.0
fluent-compiler==0.3
fluent.syntax==0.19.0
fluentogram==1.1.7
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
Jinja2==3.1.4
magic-filter==1.0.12
Mako==1.3.6
MarkupSafe==3.0.1
marshmallow==3.22.0
multidict==6.1.0
ordered-set==4.1.0
packaging==24.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
python-dotenv==1.0.1
pytz==2024.2
redis==5.1.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
watchdog==2.3.1
yarl==1.12.1
pre-commit
ruff
black
mypy
pytest
pytest-asyncio
aiosqlite
asyncpg
redis>=5.0.0
